+++ packages/client-proxy/internal/edge/handlers/sandbox-delete.go
@@ -10,6 +10,7 @@ import (
 	"github.com/e2b-dev/infra/packages/proxy/internal/edge/sandboxes"
 	grpcorchestrator "github.com/e2b-dev/infra/packages/shared/pkg/grpc/orchestrator"
 	"github.com/e2b-dev/infra/packages/shared/pkg/http/edge"
+	l "github.com/e2b-dev/infra/packages/shared/pkg/logger"
 	"github.com/e2b-dev/infra/packages/shared/pkg/telemetry"
 )
 
@@ -47,6 +48,6 @@ func (a *APIStore) V1DeleteSandbox(c *gin.Context, sandboxId api.SandboxId) {
 		return
 	}
 
-	zap.L().Info("Sandbox deleted", zap.String("sandbox_id", sandboxId))
+	zap.L().Info("Sandbox deleted", l.WithSandboxID(sandboxId))
 	c.Status(http.StatusOK)
 }
